ReleaseCandidate_0.4, minor changes because of Clang-Tidy Recommendations

+++ BigUnsignInt.cpp
@@ -23,21 +23,21 @@ BigUnsignInt &operator*=(BigUnsignInt &bigIntFactor, const unsigned long int &lo
         }
 
     bigIntFactor.digits.resize(interimProduct.size());
-    for (unsigned short int s, i = 0, t = 0; i < interimProduct.size(); i++) {
+    for (unsigned long long int s, i = 0, t = 0; i < interimProduct.size(); i++) {
         s = t + interimProduct[i];
         interimProduct[i] = s % 10;
         t = s / 10;
         bigIntFactor.digits[i] = interimProduct[i];
     }
 
-    for (int i = interimProduct.size() - 1; i >= 1 && !interimProduct[i]; i--) {
+    for (unsigned long long int i = interimProduct.size() - 1; i >= 1 && !interimProduct[i]; i--) {
         bigIntFactor.digits.pop_back();
     }
 
     return bigIntFactor;
 }
 std::ostream &operator<<(std::ostream &out,const BigUnsignInt &numberToPrint){
-    for (int i = numberToPrint.digits.size() - 1; i >= 0; i--) {
+    for (unsigned long long int i = numberToPrint.digits.size() - 1; i >= 0; i--) {
         std::cout << (unsigned short int)numberToPrint.digits[i];}
     return std::cout;
 }
